# Fix issue looking up routes that don't exist