--- a/test/issue-59.test.js
+++ b/test/issue-59.test.js
@@ -24,3 +24,14 @@ test('multi-character prefix', t => {
 
   t.equal(findMyWay.find('GET', '/bulk'), null)
 })
+
+test('with parameter', t => {
+  t.plan(2)
+  const findMyWay = FindMyWay()
+
+  findMyWay.on('GET', '/bb/:foo/aa/', noop)
+  findMyWay.on('GET', '/bb/:foo/aa/bulk', noop)
+
+  t.equal(findMyWay.find('GET', '/bulk'), null)
+  t.equal(findMyWay.find('GET', '/bb/foo/bulk'), null)
+})