How this fix works:

When matching a path to a prefix in a `node`, this part of the code finds the longest common prefix. Before it would only look for the next part of the path if the path matched the entire prefix, otherwise it would continue matching the same path. This was wrong because then this would happen:

```
path = 'bulk'
prefix = 'b/'
commonPrefix = 'b'
-- does not update path --
finds child with prefix = 'bulk'
goes back to the top and matches 'bulk'
return handler for '/b/bulk'
```

With this fix it does this:

```
path = 'bulk'
prefix = 'b/'
commonPrefix = 'b'
update path = 'ulk'
no children match 'ulk'
return null
```

<details>
<summary>benchmarks</summary>

**before**
```
lookup static route x 26,167,641 ops/sec ±1.42% (89 runs sampled)
lookup dynamic route x 1,568,707 ops/sec ±0.56% (93 runs sampled)
lookup dynamic multi-parametric route x 872,150 ops/sec ±1.18% (90 runs sampled)
lookup dynamic multi-parametric route with regex x 800,702 ops/sec ±1.98% (83 runs sampled)
lookup long static route x 2,584,658 ops/sec ±1.07% (92 runs sampled)
find static route x 32,397,375 ops/sec ±1.69% (89 runs sampled)
find dynamic route x 1,942,803 ops/sec ±1.57% (88 runs sampled)
find dynamic multi-parametric route x 1,066,697 ops/sec ±1.29% (88 runs sampled)
find dynamic multi-parametric route with regex x 926,740 ops/sec ±1.48% (90 runs sampled)
find long static route x 4,905,622 ops/sec ±1.39% (89 runs sampled)
```

**after**
```
lookup static route x 27,301,612 ops/sec ±1.26% (88 runs sampled)
lookup dynamic route x 1,576,759 ops/sec ±0.55% (94 runs sampled)
lookup dynamic multi-parametric route x 890,252 ops/sec ±0.56% (94 runs sampled)
lookup dynamic multi-parametric route with regex x 845,854 ops/sec ±0.43% (93 runs sampled)
lookup long static route x 2,425,642 ops/sec ±0.64% (93 runs sampled)
find static route x 30,674,123 ops/sec ±0.81% (88 runs sampled)
find dynamic route x 1,875,931 ops/sec ±0.76% (93 runs sampled)
find dynamic multi-parametric route x 1,037,566 ops/sec ±0.63% (92 runs sampled)
find dynamic multi-parametric route with regex x 955,495 ops/sec ±0.90% (94 runs sampled)
find long static route x 4,807,380 ops/sec ±0.67% (94 runs sampled)
```
</details>
<br>

Alternative to #61.

Fixes #59
Fixes #60