@nwoltman 2018-03-02
@delvedor I added that test. I think I don't understand the code well enough to fix it.