@delvedor 2018-03-02
Add `if ((isStatic === true && len > 0) || len === prefixLen) {` fixes everything, but there is a small regression in the performances.
Apparently if we are working with a static route the length of the common prefix just need to be higher than 0, while if we are working with a dynamic route it **must** be the same of the prefix length.