@delvedor 2018-03-02
I've fixed the test.
I'll open another pr with some optimization that will remove the little regression that we are having here.