@delvedor 2018-03-02
Landed in `v1.10.2`, thanks!