@nwoltman 2018-03-02
So it turns out that my idea with `len > 0` wasn't good enough. With `v1.10.2` this case fails:

```js
var router = require('find-my-way')()

router.on('GET', '/b/', function b () {})
router.on('GET', '/b/bulk', function b_bulk () {})
router.on('GET', '/b/ulk', function b_ulk () {})

console.log(router.find('GET', '/bulk')) // { handler: [Function: b_ulk], ... }
```