@delvedor 2018-03-02
@nwoltman that's weird!
Since is a very edge case I'll consider this bug as fixed, let's open another issue for the test case you have just found :)